feat: mutual links (#339)
* build: update misskey_dart

* feat: mutual links

+++ lib/view/page/list/lists_page.dart
@@ -63,7 +63,7 @@ class ListsPage extends ConsumerWidget {
                                             :final userEachUserListsLimit?,
                                           ))
                                         ' / ${NumberFormat().format(
-                                          userEachUserListsLimit.toInt(),
+                                          userEachUserListsLimit,
                                         )}',
                                     ].join(),
                                   ),
